% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opuslab-check-dirs-files.R
\name{struct_data_metadata}
\alias{struct_data_metadata}
\title{Check data and metadata files generated by Bruker OPUS lab spectral 
measurement workflow}
\usage{
struct_data_metadata(data_root, pattern = "\\\\.ods$")
}
\description{
Perform folder, file naming convention, and data and metadata
integrity tests.
}
