% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-data-metadata.R
\name{test_data_metadata}
\alias{test_data_metadata}
\title{Check data and metadata files generated by Bruker OPUS lab spectral 
measurement workflow}
\usage{
test_data_metadata(data_root, pattern = "\\\\.ods$")
}
\description{
Perform folder, file naming convention, and data and metadata
integrity tests.
}
